// src/interfaces/enums/gastos.enums.ts
export enum CategoriaGasto {
  ALIMENTACION = "ALIMENTACION",
  MEDICAMENTOS = "MEDICAMENTOS",
  COMPRA_ANIMAL = "COMPRA_ANIMAL",
  MANO_OBRA = "MANO_OBRA",
  INSUMOS_AGRICOLAS = "INSUMOS_AGRICOLAS",
  MANTENIMIENTO = "MANTENIMIENTO",
  TRANSPORTE = "TRANSPORTE",
  SERVICIOS_PUBLICOS = "SERVICIOS_PUBLICOS",
  IMPUESTOS = "IMPUESTOS",
  SEGUROS = "SEGUROS",
  EQUIPOS = "EQUIPOS",
  INFRAESTRUCTURA = "INFRAESTRUCTURA",
  VETERINARIA = "VETERINARIA",
  REPRODUCCION = "REPRODUCCION",
  OTROS = "OTROS",
}

export enum MetodoPago {
  EFECTIVO = "EFECTIVO",
  TRANSFERENCIA = "TRANSFERENCIA",
  TARJETA_CREDITO = "TARJETA_CREDITO",
  TARJETA_DEBITO = "TARJETA_DEBITO",
  CHEQUE = "CHEQUE",
  OTRO = "OTRO",
}