// src/interfaces/enums/ingresos.enums.ts
export enum CategoriaIngreso {
  VENTA_ANIMAL = "VENTA_ANIMAL",
  VENTA_LECHE = "VENTA_LECHE",
  VENTA_CARNE = "VENTA_CARNE",
  VENTA_CUERO = "VENTA_CUERO",
  VENTA_LANA = "VENTA_LANA",
  VENTA_HUEVOS = "VENTA_HUEVOS",
  VENTA_ESTIERCOL = "VENTA_ESTIERCOL",
  SUBSIDIOS = "SUBSIDIOS",
  SEGUROS = "SEGUROS",
  ARRENDAMIENTOS = "ARRENDAMIENTOS",
  SERVICIOS = "SERVICIOS",
  REPRODUCCION = "REPRODUCCION",
  OTROS = "OTROS",
}